Updated impute with special category for categorical

+++ NullPredictor.py
@@ -69,7 +69,7 @@ class NullPredictor():
     		if self.target_transformer[col] != None:
     			predicted = self.target_transformer[col].inverse_transform(predicted)
 
-    		data[col].iloc[null_idx] = predicted.astype('int')
+    		data[col].iloc[null_idx] = predicted
 
     	return data
